Fix import pthonjsonlogger.jsonlogger
Fixes: #29

Although efforts were made for backwards compatibility using
`__getattr__` in `pythonjsonlogger/__init__.py`, this only worked
if you were doing `from pythonjsonlogger import jsonlogger`. When
importing by the full name `import pythonjsonlogger.jsonlogger` it
it would fail to locate the module file and thus not be able to
produce a module spec which then of course causes the import to fail.

We get around this by actually having a module that imports the names
it needs from the new locations.